requirements.txt
requests
beautifulsoup4
lxml
scrapy
anthropic
sentence-transformers
transformers
torch
faiss-cpu
langchain
langchain-community
langchain-huggingface
nltk
scikit-learn
pandas
numpy
tqdm
python-dotenv
jupyter
matplotlib
seaborn